backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
openai==1.16.2
chromadb==0.4.24
watchdog==4.0.0
PyAudio==0.2.14
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
pydantic==2.6.4
sqlalchemy==2.0.29
python-dotenv==1.0.1
python-multipart==0.0.9